Provides almost complete abstract implementations.
