+++ groovy-language-server/src/main/java/com/palantir/ls/DefaultLanguageServerState.java
@@ -2,7 +2,7 @@
  * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
+ * you may not use this file except in compliance ctes the License.
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
@@ -14,20 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.ls.services;
+package com.palantir.ls;
 
-import com.palantir.ls.api.LanguageServerState;
-import io.typefox.lsapi.DidChangeConfigurationParams;
-
-public final class DefaultWorkspaceService extends AbstractWorkspaceService {
-    private final LanguageServerState state;
-
-    public DefaultWorkspaceService(LanguageServerState state) {
-        this.state = state;
-    }
-
-    @Override
-    protected LanguageServerState getState() {
-        return state;
-    }
+public class DefaultLanguageServerState extends AbstractLanguageServerState {
 }
